+++ lib/ecs-kingpin/aws.rb
