+++ lib/ecs-kingpin/log.rb
