+++ lib/ecs-kingpin/parse.rb
