+++ lib/ecs-kingpin/service.rb
