+++ lib/ecs-kingpin/task.rb
